Moved Teams.js from components folder to views folder as this is an actual page view
