+++ client/src/views/Teams.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import Navbar from './Navbar';
+import Navbar from '../components/Navbar';
 
 function Teams() {
   const [teams, setTeams] = useState([]);
